Fix memory leak preventing caching methods.
