+++ wots.c
@@ -6,9 +6,258 @@
 #include "wots.h"
 #include "hash_address.h"
 #include "params.h"
+//#ifdef CONSTANTSUM
+//#include "constant_sum.h"
+//#endif
+
+
 #ifdef CONSTANTSUM
-#include "constant_sum.h"
+
+
+/**
+ * Binomial coefficient n choose k for arbitrary precisions
+ * integers, using GMP. Do not use uint, since it
+ * may be called with negative values and return 0.
+ */
+static void binomial(int32_t n, int32_t k, mpz_t out)
+{
+	mpz_init(out);
+	if(n < k || n<0 || k<0)
+		return;
+	mpz_bin_uiui(out, n, k);
+}
+
+/**
+ * Ranking function, returns the j-th rank for params,
+ * used to find bounds in the constant-sum encoding.
+ */
+static void rank(int32_t t, int32_t n, int32_t s, int32_t j, mpz_t out) {
+	mpz_init(out);
+	int32_t aux = s/(n+1); //Floor of sum/(max+1)
+	int32_t k = (t < aux)? t : aux; //min(blocks,aux)
+    int32_t i;
+	mpz_t a; mpz_init(a);
+	mpz_t b; mpz_init(b);
+	mpz_t c; mpz_init(c);
+	for(i = 0; i <= k; i++ ) { 
+		binomial(t-1,i,a);
+		binomial(s - (n+1)*i + t-1, t-1, b);
+		binomial(s - (n+1)*i + t -2 -j, t-1, c);
+		mpz_sub(b,b,c);
+		if(i%2==0){
+			mpz_addmul(out, a,b);
+		}else{
+			mpz_submul(out, a,b);
+		}
+	}
+	mpz_clear(a);
+	mpz_clear(b);
+	mpz_clear(c);
+}
+
+#if defined(BCACHED) || defined(VCACHED)
+mpz_t bcache[T][S+1][N+1];
+void load_bcache(const int32_t t, const int32_t n, const int32_t s)
+{
+	int b,z,j;
+	for(b = 0; b < t; b++) {
+		for(z = 0; z<=s; z++){
+			#pragma omp parallel for
+			for(j = 0; j<= n; j++){
+				if(j<=z){
+					rank(b+1,n,z,j,bcache[b][z][j]);
+					//rank(b+1,n,z,j,*(*(*(bcache +b)+z)+j));
+				}
+			}
+		}
+	}
+
+}
 #endif
+
+/**
+ * Given a constant-sum encoding and the original digest, verifies
+ * if encoding corresponds to the digest, similar to a checksum.
+ */
+int check_encoding(mpz_t I, int32_t t, int32_t n, int32_t s, 
+						  int *encoding) 
+{
+	mpz_t left; mpz_init(left);
+	mpz_t right; mpz_init(right);
+	int k = 0;
+	for(int b = 0; b <t; b++) {
+		k = encoding[t-b-1];
+#ifdef VCACHED
+		(void)n;
+		if(k == 0) {
+			mpz_init(left);
+		} else {
+			mpz_set(left, bcache[t-b-1][s][k-1]);
+		}
+		mpz_set(right, bcache[t-b-1][s][k]);
+#else
+		rank(t-b, n, s, k-1, left);
+		rank(t-b, n, s, k, right);
+#endif
+		if( mpz_cmp(left,I)>0 ) { 
+			mpz_clear(left);
+			mpz_clear(right);
+			return -1; 
+		}
+		if( mpz_cmp(right,I)<=0 ) { 
+			mpz_clear(left);
+			mpz_clear(right);
+			return -1; 
+		}
+		s -= encoding[t-b-1];
+		mpz_sub(I,I,left);
+	}
+	mpz_clear(left);
+	mpz_clear(right);
+	return 0;
+}
+
+#ifdef BINARYSEARCH
+
+/**
+ * Constan-sum encoding function using binary search. Encodes an 
+ * integer 0 <= i <=2^m into t partitions of integers from 0 to n 
+ * and summing s. Writes the encoding into output.
+ */
+void toConstantSum(mpz_t I, int32_t t, int32_t n, int32_t s, 
+						  int *output) 
+{
+	int count, k, it, step,b;
+	mpz_t keep; mpz_init(keep);
+	for (b = 0; b < t; b++ ) {
+		count = ( n < s)?n:s;
+		k = 0;
+		while ( count > 0 ) {
+			it = k;
+			step = count/2;
+			it = it + step;
+#ifdef BCACHED
+			//mpz_set(keep, *(*(*(bcache +t-b-1)+s)+it));
+			mpz_set(keep, bcache[t-b-1][s][it]);
+#else
+			rank(t-b, n, s, it, keep);
+#endif
+			if ( mpz_cmp(I,keep)>=0) {
+				k = ++it;
+				count -= step + 1;
+			} else {
+				count = step;
+			}
+		}
+		output[t-b-1] = k;
+		if(k>0) {
+			rank(t-b, n, s, k-1, keep);
+			mpz_sub(I,I,keep);
+			s -= k;
+		}
+	}
+	mpz_clear(keep);
+}
+
+#else
+/**
+ * Computes cardinality of the set of t-tuples with values ranging
+ * from 0 to n that sum s.
+ */
+static void constantSumLen(const int32_t t, const int32_t n, const int32_t s, 
+						   mpz_t out) 
+{
+	mpz_init(out);
+	int32_t aux = s/(n+1); //Floor of sum/(max+1)
+	int32_t k = (t < aux)? t : aux; //min(blocks,aux)
+    int32_t i;
+	mpz_t a;
+	mpz_t b;
+	for(i = 0; i <= k; i++ ) {
+		binomial(t,i,a);
+		binomial(s-(n+1)*i+t-1,t-1, b);
+		if(i%2==0){
+			mpz_addmul(out, a,b);
+		}else{
+			mpz_submul(out, a,b);
+		}
+	}
+			mpz_clear(a);
+			mpz_clear(b);
+}
+
+
+#ifdef CACHED
+mpz_t cache[T-1][S+1];
+void load_cache(const int32_t t, const int32_t n, const int32_t s)
+{
+	int b,z;
+	for(b = 1; b < t; b++) {
+		for(z = 0; z<=s; z++){
+			constantSumLen(b,n,z,cache[b-1][z]);
+		}
+	}
+
+}
+#endif
+
+/**
+ * Constan-sum encoding function. Encodes an integer 0 <= i <=2^m
+ * into t partitions of integers from 0 to n and summing s. Writes
+ * the encoding into output.
+ */
+void toConstantSum(mpz_t I, int32_t t, int32_t n, int32_t s, 
+						  int *output) 
+{
+	if (t == 1){
+		output[0] = s;
+		return;
+	}
+	int32_t k = 0;
+	mpz_t aux; mpz_init(aux);
+	mpz_t left;	mpz_init(left);
+	mpz_t right; mpz_init(right);
+	constantSumLen(t - 1, n, s, right);
+	while ( !( mpz_cmp(I,left)>=0 && mpz_cmp(I,right)<0) ) {
+		k++; 
+		mpz_set(left,right);
+#ifdef CACHED
+		mpz_set(aux, cache[t-2][s-k]);
+#else
+		constantSumLen(t-1,n,s-k,aux);
+#endif
+		mpz_add(right,right,aux);
+	}
+	output[t-1] = k;
+	mpz_sub(I,I,left);
+	mpz_clear(aux);
+	mpz_clear(left);
+	mpz_clear(right);
+	toConstantSum(I, t - 1, n, s-k, output);
+}
+
+#endif
+#endif
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 
 /**
  * Helper method for pseudorandom key generation.
@@ -151,6 +400,7 @@ void wots_sign(const xmss_params *params,
 #ifdef CONSTANTSUM
 	mpz_t I; mpz_init(I); mpz_import(I, params->n,1,1,0,0, msg);
 	toConstantSum(I, params->wots_len, params->wots_w, params->wots_s, lengths);
+	mpz_clear(I);
 #ifdef VERIFY
 		short aux;
 #endif
@@ -198,10 +448,13 @@ void wots_pk_from_sig(const xmss_params *params, unsigned char *pk,
 	}
 
 	if(check_encoding(I, params->wots_len, params->wots_w, params->wots_s, lengths)){
+		mpz_clear(I);
 		return;
 	}
+	mpz_clear(I);
 #else
 	toConstantSum(I, params->wots_len, params->wots_w, params->wots_s, lengths);
+	mpz_clear(I);
 #endif
 #else
     chain_lengths(params, lengths, msg);
